# PsychoMedikApp/PsychoMedikApp/OpenAPIs/swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "PsychoMedikAPI",
    "version": "1.0"
  },
  "paths": {
...
    "/api/Pacjent": {
      "get": {
        "tags": [
          "Pacjent"
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Pacjent"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Pacjent"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Pacjent"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Pacjent"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Pacjent"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/Pacjent"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/Pacjent"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/Pacjent"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Pacjent"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/Pacjent"
                }
              }
            }
          }
        }
      }
    },
    "/api/Pacjent/{id}": {
      "get": {
        "tags": [
          "Pacjent"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/Pacjent"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Pacjent"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/Pacjent"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Pacjent"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Pacjent"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/Pacjent"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/Pacjent"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "delete": {
        "tags": [
          "Pacjent"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
...
    "/api/Pracownik": {
      "get": {
        "tags": [
          "Pracownik"
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Pracownik"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Pracownik"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Pracownik"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Pracownik"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Pracownik"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/Pracownik"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/Pracownik"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/Pracownik"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Pracownik"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/Pracownik"
                }
              }
            }
          }
        }
      }
    },
    "/api/Pracownik/{id}": {
      "get": {
        "tags": [
          "Pracownik"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/Pracownik"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Pracownik"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/Pracownik"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Pracownik"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Pracownik"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/Pracownik"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/Pracownik"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "delete": {
        "tags": [
          "Pracownik"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
...
    "/api/Wizyta": {
      "get": {
        "tags": [
          "Wizyta"
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Wizyta"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Wizyta"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Wizyta"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Wizyta"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Wizyta"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/Wizyta"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/Wizyta"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/Wizyta"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Wizyta"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/Wizyta"
                }
              }
            }
          }
        }
      }
    },
    "/api/Wizyta/{id}": {
      "get": {
        "tags": [
          "Wizyta"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/Wizyta"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Wizyta"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/Wizyta"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Wizyta"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Wizyta"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/Wizyta"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/Wizyta"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "delete": {
        "tags": [
          "Wizyta"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "Choroba": {
        "required": [
          "nazwa"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "czyAktywny": {
            "type": "boolean"
          },
          "dataUtworzenia": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "dataModyfikacji": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "opis": {
            "type": "string",
            "nullable": true
          },
          "nazwa": {
            "type": "string"
          },
          "objawy": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Objaw"
            },
            "nullable": true
          },
          "historiaChorob": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/HistoriaChoroby"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "Harmonogram": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "czyAktywny": {
            "type": "boolean"
          },
          "dataUtworzenia": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "dataModyfikacji": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "idPracownika": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "pracownik": {
            "$ref": "#/components/schemas/Pracownik"
          },
          "dzienRoku": {
            "type": "integer",
            "format": "int32"
          },
          "godzinaRozpoczecia": {
            "type": "integer",
            "format": "int32"
          },
          "godzinaZakonczenia": {
            "type": "integer",
            "format": "int32"
          }
        },
        "additionalProperties": false
      },
      "HistoriaChoroby": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "czyAktywny": {
            "type": "boolean"
          },
          "dataUtworzenia": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "dataModyfikacji": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "opis": {
            "type": "string",
            "nullable": true
          },
          "idPacjenta": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "pacjent": {
            "$ref": "#/components/schemas/Pacjent"
          },
          "idChoroby": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "choroba": {
            "$ref": "#/components/schemas/Choroba"
          },
          "idPracownika": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "pracownik": {
            "$ref": "#/components/schemas/Pracownik"
          },
          "dataZdiagnozowania": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "dataWyleczenia": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "Objaw": {
        "required": [
          "nazwa"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "czyAktywny": {
            "type": "boolean"
          },
          "dataUtworzenia": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "dataModyfikacji": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "opis": {
            "type": "string",
            "nullable": true
          },
          "nazwa": {
            "type": "string"
          },
          "choroby": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Choroba"
            },
            "nullable": true
          },
          "pacjenci": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Pacjent"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "Pacjent": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "czyAktywny": {
            "type": "boolean"
          },
          "dataUtworzenia": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "dataModyfikacji": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "opis": {
            "type": "string",
            "nullable": true
          },
          "imie": {
            "type": "string",
            "nullable": true
          },
          "nazwisko": {
            "type": "string",
            "nullable": true
          },
          "dataUrodzenia": {
            "type": "string",
            "format": "date-time"
          },
          "plec": {
            "type": "boolean"
          },
          "idPracownika": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "pracownik": {
            "$ref": "#/components/schemas/Pracownik"
          },
          "objawy": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Objaw"
            },
            "nullable": true
          },
          "historiaChoroby": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/HistoriaChoroby"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "Pracownik": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "czyAktywny": {
            "type": "boolean"
          },
          "dataUtworzenia": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "dataModyfikacji": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "opis": {
            "type": "string",
            "nullable": true
          },
          "imie": {
            "type": "string",
            "nullable": true
          },
          "nazwisko": {
            "type": "string",
            "nullable": true
          },
          "dataUrodzenia": {
            "type": "string",
            "format": "date-time"
          },
          "dataZatrudnienia": {
            "type": "string",
            "format": "date-time"
          },
          "dataRezygnacji": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "plec": {
            "type": "boolean"
          },
          "stanCywilny": {
            "type": "string",
            "nullable": true
          },
          "idStanowiska": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "stanowisko": {
            "$ref": "#/components/schemas/Stanowisko"
          },
          "pacjenci": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Pacjent"
            },
            "nullable": true
          },
          "harmonogram": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Harmonogram"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "Stanowisko": {
        "required": [
          "nazwa"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "czyAktywny": {
            "type": "boolean"
          },
          "dataUtworzenia": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "dataModyfikacji": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "opis": {
            "type": "string",
            "nullable": true
          },
          "nazwa": {
            "type": "string"
          },
          "pracownicy": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Pracownik"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "Wizyta": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "czyAktywny": {
            "type": "boolean"
          },
          "dataUtworzenia": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "dataModyfikacji": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "opis": {
            "type": "string",
            "nullable": true
          },
          "dataWizyty": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "idPacjenta": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "pacjent": {
            "$ref": "#/components/schemas/Pacjent"
          },
          "idLekarza": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "pracownik": {
            "$ref": "#/components/schemas/Pracownik"
          }
        },
        "additionalProperties": false
      }
    }
  }
}